+++ examples/pingpong/src/pong.rs
@@ -44,7 +44,7 @@ fn ping_received(ctx: &ArcContext<AgentProps>, message: &Message) -> Result<(), 
 	Ok(())
 }
 
-#[tokio::main]
+#[tokio::main(flavor="current_thread")]
 async fn main() -> Result<(), DimasError> {
 	// a tracing subscriber writing logs
 	tracing_subscriber::fmt::init();